api/requirements.txt
Django==3.2.8
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
PyJWT==2.3.0
drf-yasg==1.20.0
django-filter==21.1
psycopg2-binary==2.9.1
django-cors-headers==3.10.0
djongo==1.3.6
pymongo[srv]
